requirements.txt
# Install system dependency first:
#   sudo apt-get install tesseract-ocr

opencv-python>=4.8.0
pytesseract>=0.3.10
paho-mqtt>=2.0.0
PyYAML>=6.0.1
